% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Get CoronaNet Event Data}
\usage{
get_event(
  countries = "All",
  type = "All",
  type_sub_cat = "All",
  from = "2019-12-31",
  to = "2027-07-01"
)
}
\arguments{
\item{countries}{A character vector of country name(s), e.g., c("Yemen", "Saudi Arabia"). "All" is used as the default.}

\item{type}{A character vector of policy types, e.g., c("Lockdown", "Curfew"). "All" is used as the default. See https://www.coronanet-project.org/taxonomy.html? for a list of policy types.}

\item{type_sub_cat}{A character vector of policy types, e.g., c("Self-testing", "Drive-in testing centers"). "All" is used as the default. See https://www.coronanet-project.org/taxonomy.html? for a list of policy subtypes and their related policy types.}

\item{from}{A character vector for the earliest start date, e.g., "2019-12-31".}

\item{to}{A character vector for the last end date, e.g., "2019-06-01".}
}
\value{
A dataframe
}
\description{
Get CoronaNet Event Data
}
\examples{
get_event(countries = "All", type = "All", type_sub_cat = "All", from = "2019-12-31", to = "2027-07-01")
}
